Add files via upload
Write a program in C++ using File Handle.
in this program the user will enter the name of two files. the first file will be considered as source which means it is to be checked with the second file. the system will open both file in text mode and compare their assimilities ignoring up and lower case. after checking the program will be give a summary has flow.  
1. the total number of rewards in the first file. 
 2. total number of wards in the second file. 
 3. percentage similarity based on the number of words found.